boj/p5k/p5500/p5582_공통_부분_문자열_test.go
package p5500

import (
	"bufio"
	"fmt"
	"strings"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestSolve5582(t *testing.T) {
	t.Log("https://www.acmicpc.net/problem/5582")

	//goland:noinspection SpellCheckingInspection
	tests := []struct {
		give string
		want string
	}{
		{
			`ABRACADABRA
ECADADABRBCRDARA`,
			"5",
		},
		{
			`UPWJCIRUCAXIIRGL
SBQNYBSBZDFNEV`,
			"0",
		},
	}

	for i, tt := range tests {
		t.Run(fmt.Sprintf("%d", i), func(t *testing.T) {
			reader := bufio.NewReader(strings.NewReader(tt.give))
			buf := new(strings.Builder)
			writer := bufio.NewWriter(buf)

			Solve5582(reader, writer)

			err := writer.Flush()
			assert.NoError(t, err)

			got := buf.String()
			assert.Equal(t, tt.want, got)
		})
	}
}
